feat: moved components folder of Home page
