+++ src/components/Home/Bar.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
 // utils
-import { BarStatus } from '../../../utils'
+import { BarStatus } from '../../utils'
 
 interface IProps {
   value: number
