+++ src/components/Home/BottomPanel.tsx
