+++ src/components/Home/Legend.tsx
