+++ src/components/Home/SidePanel.tsx
@@ -2,45 +2,48 @@ import React, { useState } from 'react'
 
 // components
 import Highlight from 'react-highlight.js'
-import { Header, Dropdown } from '../../../shared'
+import { Header, Dropdown } from '../../shared'
 
-const bubbleSortCode = `let bubbleSort = (inputArr) => {
-	let len = inputArr.length;
-	let swapped;
-	do {
-		swapped = false;
-		for (let i = 0; i < len; i++) {
-			if (inputArr[i] > inputArr[i + 1]) {
-				let tmp = inputArr[i];
-				inputArr[i] = inputArr[i + 1];
-				inputArr[i + 1] = tmp;
-				swapped = true;
-			}
-		}
-	} while (swapped);
-	do {
-		swapped = false;
-		for (let i = 0; i < len; i++) {
-			if (inputArr[i] > inputArr[i + 1]) {
-				let tmp = inputArr[i];
-				inputArr[i] = inputArr[i + 1];
-				inputArr[i + 1] = tmp;
-				swapped = true;
-			}
-		}
-	} while (swapped);
-	do {
-		swapped = false;
-		for (let i = 0; i < len; i++) {
-			if (inputArr[i] > inputArr[i + 1]) {
-				let tmp = inputArr[i];
-				inputArr[i] = inputArr[i + 1];
-				inputArr[i + 1] = tmp;
-				swapped = true;
-			}
-		}
-	} while (swapped);
-	return inputArr;
+const bubbleSortCode = `
+let bubbleSort = (inputArr) => {
+  let len = inputArr.length;
+  let swapped;
+  do {
+    swapped = false;
+    for (let i = 0; i < len; i++) {
+      if (inputArr[i] > inputArr[i + 1]) {
+        let tmp = inputArr[i];
+        inputArr[i] = inputArr[i + 1];
+        inputArr[i + 1] = tmp;
+        swapped = true;
+      }
+    }
+  } while (swapped);
+
+  do {
+    swapped = false;
+    for (let i = 0; i < len; i++) {
+      if (inputArr[i] > inputArr[i + 1]) {
+        let tmp = inputArr[i];
+        inputArr[i] = inputArr[i + 1];
+        inputArr[i + 1] = tmp;
+        swapped = true;
+      }
+    }
+  } while (swapped);
+  
+  do {
+    swapped = false;
+    for (let i = 0; i < len; i++) {
+      if (inputArr[i] > inputArr[i + 1]) {
+        let tmp = inputArr[i];
+        inputArr[i] = inputArr[i + 1];
+        inputArr[i + 1] = tmp;
+        swapped = true;
+      }
+    }
+  } while (swapped);
+  return inputArr;
 };`
 
 const algorithmsOptions = [
